Refactor: Fix table name in subjectListRouter query
The code changes in the `subjectList.ts` file fix a table name typo in the `subjectListRouter` query. The table name is updated from "subjectList" to "SubjectList" to ensure correct data retrieval. This refactor improves the accuracy and reliability of the query.

Note: This commit message is based on the code changes provided.