Refactor parseTimeString method to improve time parsing
The code changes in the `interfaces.ts` file add a new method `parseTimeString` to the `SchedulerClass` interface. This method is responsible for parsing a string representation of time and returning a `Date` object. This improvement enhances the functionality of the scheduling feature by providing a standardized way to parse time inputs.